refactor(TimezoneSelector): streamline timezone option retrieval by utilizing timezoneOptionsCache and remove redundant functions
